Add files via upload
This project demonstrates the fundamentals of REST API development using FLASK in python. The API manages user data stored in an in-memory dictionary, allowing full CRUD operations.
---->POST: Add a new user with a unique ID and name.
---->GET: Retrieve all users or a specific user by ID.
---->PUT: Update an existing user's details.
---->DELETE: Remove a user by ID.
         The project emphasizes core API concepts such as handling JSON requests/responses, returning proper HTTP status codes, and managing error cases(e.g., missing data or non-existing users).
Developers can interact with the API using POSTMAN or CURL to simulate client requests. This setup serves as a hands-on introduction to backend development, showing how to design, implement, and test RESTful services in a simple yet practical way.